% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{CheckSufficientPower}
\alias{CheckSufficientPower}
\title{Check sufficiant power of all possible covariate groups}
\usage{
CheckSufficientPower(metaFile, nnodes = 1, cutoff = 5)
}
\arguments{
\item{metaFile}{a tab delimited file or data frame with row and colum names listing metadata for all samples}

\item{nnodes}{number of nodes to be used for parallel computing}

\item{cutoff}{minimamal number of sample size for each covariate in order to have sufficient power, default = 5}
}
\value{
data frame containing sample sizes for all covariates and sufficientPower as yes(1)/no(0)
}
\description{
Check sufficiant power of all possible covariate groups
}
