% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportLongPrior.R
\name{ImportLongPrior}
\alias{ImportLongPrior}
\title{ImportLongPrior}
\usage{
ImportLongPrior(longPrior, featureMat, metaMat)
}
\arguments{
\item{longPrior}{long-format dataframe as generated by
Metadeconfound(returnLong = TRUE). Must contain at least one column
containing feature names and one column containing associated metadata names,
called "feature" and "metadata", respectively. Only associations between
features and metadata present in featureMat and metaMat will be returned.
Additionally, "Qs" and "status" (as produced by
\link[metadeconfoundR]{MetaDeconfound})columns can be supplied and will
be parsed as well.
If only "feature" and "metadata"columns are supplied, all listed
associations are assumed to be significant.
If "status" is supplied, only non-"NS" labeled associations will be kept.}

\item{featureMat}{omics features to be analyzed by \link[metadeconfoundR]{MetaDeconfound}}

\item{metaMat}{metadata to be analyzed by \link[metadeconfoundR]{MetaDeconfound}}
}
\value{
wide-format dataframe that can be used as minQValues parameter in
\link[metadeconfoundR]{MetaDeconfound}
}
\description{
ImportLongPrior imports prior knowledge of associations between individual
features and metadata in form of a long-format dataframe.
}
\details{
This function is meant to facilitate incorporation of prior
knowledge about associations between measured omics features and available
metadata both from earlier metadeconfoundR runs by supplying the long-format
Metadeconfound() output directly or by supplying a simple list of known
associations from other studies.
}
\examples{
data(reduced_feature)
data(metaMatMetformin)
\donttest{
example_output <- MetaDeconfound(featureMat = reduced_feature,
                                  metaMat = metaMatMetformin,
                                  returnLong = TRUE)

minQValues <- ImportLongPrior(longPrior = example_output,
                                featureMat = reduced_feature,
                                metaMat = metaMatMetformin,
                                  )
}

example_output <- MetaDeconfound(featureMat = reduced_feature,
                                  metaMat = metaMatMetformin,
                                  minQValues = minQValues)
}
