% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDeconfound.R
\name{MetaDeconfound}
\alias{MetaDeconfound}
\title{metadeconfoundR}
\usage{
MetaDeconfound(featureMat, metaMat, nnodes = 1, adjustMethod = "fdr",
  robustCutoff = 5, QCutoff = 0.1, DCutoff = 0, PHS_cutoff = 0.05,
  NA_imputation = "remove", ...)
}
\arguments{
\item{featureMat}{a data frame with row(sample ID)
and column(feature such as metabolite or microbial OTU )
names, listing features for all samples}

\item{metaMat}{a data frame with row(sample ID) and
column(meta data such as age,BMI and all possible confounders)
names listing metadata for all samples. first column should be case status
with case=1 and control=0.}

\item{nnodes}{number of nodes/cores to be used for parallel processing}

\item{adjustMethod}{multiple testing p-value correction using one of the
methods of stats::p.adjust.methods}

\item{robustCutoff}{minimamal number of sample size for each covariate
in order to have sufficient power for association testing}

\item{QCutoff}{significance cutoff for q-value, DEFAULT = 0.1}

\item{DCutoff}{effect size cutoff
(either cliff's delta or spearman correlation test estimate), DEFAULT = 0}

\item{PHS_cutoff}{PostHoc Significance cutoff}

\item{NA_imputation}{Missing data treatment. "remove": remove NA containing
data from analysis (default).
"others": Impute NAs  using methods from packages MICE or AMELIA
(not yet implemented)}

\item{...}{for additional arguments used internally (development/debugging)}
}
\value{
list with elements Ds = effectsize,
Ps = uncorrected p-value for naive association,
Qs = multiple testing corrected p-value/fdr,
and status = confounding/mediation status for all
feature <=> covariate combinations with following categories:
(NS = not significant, SD = strictly deconfounded, LD = laxly deconfounded,
NC = no covariates, "covariate name" = confounded by this covariate)
}
\description{
metadeconfoundR checks all feature <-> covariate combinations for
counfounding effects of covariates on feature <-> effect corellation
}
\details{
for more details and explanations please see the vignette.
}
\examples{
data(reduced_feature)
data(metaMatMetformin)
\donttest{
example_output <- MetaDeconfound(featureMat = reduced_feature,
                                  metaMat = metaMatMetformin)}
}
