% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaMatMetformin_data.R
\docType{data}
\name{metaMatMetformin}
\alias{metaMatMetformin}
\title{Documentation for the metaMatMetformin RData in /data}
\description{
set of features from the metformin dataset,
containining status for 5 dfferent proprties for 753 samples
}
\keyword{data}
