% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveAssociations.R
\name{NaiveAssociation}
\alias{NaiveAssociation}
\title{Check sufficiant power of all possible covariate groups}
\usage{
NaiveAssociation(featureFile, metaFile, nnodes = 5)
}
\arguments{
\item{featureFile}{a tab delimited file or data frame with row and colum names listing features for all samples}

\item{metaFile}{a tab delimited file or data frame with row and colum names listing metadata for all samples}

\item{nnodes}{number of nodes to be used for parallel computing}
}
\value{
data frame containing p-values and effect sizes for all simple feature/covariate associations
}
\description{
Check sufficiant power of all possible covariate groups
}
