% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_feature_data.R
\docType{data}
\name{reduced_feature}
\alias{reduced_feature}
\title{Documentation for the reduced_feature RData in /data}
\description{
reduced set of features from the metformin dataset,
containining feature measurments for 753 samples
}
\keyword{data}
