% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildHeatmap.R
\name{BuildHeatmap}
\alias{BuildHeatmap}
\title{BuildHeatmap}
\usage{
BuildHeatmap(
  metaDeconfOutput,
  q_cutoff = 0.1,
  d_cutoff = 0.01,
  cuneiform = FALSE,
  coloring = 0,
  showConfounded = TRUE,
  intermedData = FALSE
)
}
\arguments{
\item{metaDeconfOutput}{output list of a metadeconfound run}

\item{q_cutoff}{optional FDR-value cutoff used to remove
low-significance entries from data}

\item{d_cutoff}{optional effect size cutoff used to remove
low effect size entries from data}

\item{cuneiform}{optional logical parameter,
plot cuneiform instead of heatmap when cuneiform = TRUE}

\item{coloring}{optional, can be 0,1,2;
0: color all tiles according to effectsize ;
1: don't color not significant tiles
2: like 1 but also don't color confounded signal tiles}

\item{showConfounded}{optional logical parameter;
set to FALSE to remove significance markers from confounded signals}

\item{intermedData}{only return intermediate data for plotting, default = FALSE}
}
\value{
ggplot2 object
}
\description{
BuildHeatmap summarizes Metadecofound output in a heatmap or cuneiform plot
}
\details{
for more details and explanations please see the package vignette.
}
\examples{
data(reduced_feature)
data(metaMatMetformin)
\donttest{
example_output <- MetaDeconfound(featureMat = reduced_feature,
                                  metaMat = metaMatMetformin)

plotObject <- BuildHeatmap(example_output)
print(plotObject)}

}
