% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaMatMetformin_data.R
\docType{data}
\name{metaMatMetformin}
\alias{metaMatMetformin}
\title{Documentation for the metaMatMetformin RData in /data}
\description{
set of features from the metformin dataset
(Forslund et al. (2015), DOI: https://doi.org/10.1038/nature15766 ),
containing status for 5 different properties for 753 samples
}
\keyword{data}
