% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDeconfound.R
\name{MultiDeconfound}
\alias{MultiDeconfound}
\title{Multi Deconfound}
\usage{
MultiDeconfound(featureMat, metaMat, nnodes = 1, robustCutoff = 5,
  adjustMethod = "fdr", QCutoff = 0.1, DCutoff = 0, ...)
}
\arguments{
\item{featureMat}{a tab delimited file or data frame with row(sample ID) and column(feature such as XXXXX) names listing features for all samples}

\item{metaMat}{a tab delimited file or data frame with row(sample ID) and column(meta data such as age,BMI and all possible confounders) names listing metadata for all samples. first column should be case status with case=1 and control=0.}

\item{nnodes}{number of nodes/cores to be used for parallel processing}

\item{robustCutoff}{minimamal number of sample size for each covariate in order to have sufficient power for association testing}
}
\value{
XXXXdata frameXXX containing information on reducability/confounding
}
\description{
Multi Deconfound checks all feature <-> covariate combinations for counfounding effects of covariates on feature <-> effect corellation
}
