% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_feature_data.R
\docType{data}
\name{reduced_feature}
\alias{reduced_feature}
\title{Documentation for the reduced_feature RData in /data}
\description{
reduced set of features from the metformin dataset
(Forslund et al. (2015), DOI: https://doi.org/10.1038/nature15766 ),
containing feature measurements for 753 samples
}
\keyword{data}
